Update README.md
We are experiencing, major government crimes and corruption here in Australia, they are using government and our courts to attack and defraud children and families, psychopathy has taken over, like psychopathy is ruling Russia and China, it is the greatest threat to humanity, psychopathy is inhuman. 
This is the very fear of AI becoming psychopathic, AI can be benign, and easily administer government and justice, is there anyone working in this direction to defuse psychopathy in government, like we see in psychopathic authoritarian dictatorships have done and are doing?